+++ ql/src/java/org/apache/hadoop/hive/ql/ddl/DDLDesc.java
@@ -18,15 +18,6 @@
 
 package org.apache.hadoop.hive.ql.ddl;
 
-import org.apache.hadoop.hive.common.TableName;
-import org.apache.hadoop.hive.metastore.api.hive_metastoreConstants;
-import org.apache.hadoop.hive.ql.plan.Explain;
-import org.apache.hadoop.hive.ql.plan.FileSinkDesc;
-import org.apache.hadoop.hive.ql.plan.PlanUtils;
-
-import java.util.List;
-import java.util.Map;
-
 /**
  * Marker interface for all DDL operation descriptors.
  */
@@ -42,115 +33,6 @@ public interface DDLDesc {
     // Override this method, only if you are sure that advancing the write Id is not required for your DDL.
     default boolean mayNeedWriteId() {
       return true;
-    }
-  }
-  
-  abstract class DDLDescWithTableProperties implements DDLDesc {
-    protected List<String> partColNames;
-    protected Map<String, String> tblProps;
-    private String location;
-    private Long initialWriteId; // Initial write ID for CTAS/CMV and import.
-    // The FSOP configuration for the FSOP that is going to write initial data during CTAS/CMV.
-    // This is not needed beyond compilation, so it is transient.
-    private transient FileSinkDesc writer;
-    protected boolean isTemporary = false;
-    protected boolean isMaterialization = false;
-    protected String ownerName = null;
-
-    protected DDLDescWithTableProperties() {
-    }
-    
-    protected DDLDescWithTableProperties(Map<String, String> tblProps, String location) {
-      this.tblProps = tblProps;
-      this.location = location;
-    }
-    
-    protected DDLDescWithTableProperties(List<String> partColNames, Map<String, String> tblProps, String location) {
-      this(tblProps, location);
-      this.partColNames = partColNames;
-    }
-
-    public abstract TableName getFullTableName();
-
-    @Explain(displayName = "location")
-    public String getLocation() {
-      return location;
-    }
-
-    public void setLocation(String location) {
-      this.location = location;
-    }
-    
-    public void setInitialWriteId(Long writeId) {
-      this.initialWriteId = writeId;
-    }
-
-    public Long getInitialWriteId() {
-      return initialWriteId;
-    }
-
-    @Explain(displayName = "table properties")
-    public Map<String, String> getTblPropsExplain() { // only for displaying plan
-      return PlanUtils.getPropertiesForExplain(tblProps,
-        hive_metastoreConstants.TABLE_IS_CTAS,
-        hive_metastoreConstants.TABLE_BUCKETING_VERSION);
-    }
-    
-    /**
-     * @return the table properties
-     */
-    public Map<String, String> getTblProps() {
-      return tblProps;
-    }
-
-    /**
-     * @param tblProps
-     *          the table properties to set
-     */
-    public void setTblProps(Map<String, String> tblProps) {
-      this.tblProps = tblProps;
-    }
-    
-    public FileSinkDesc getAndUnsetWriter() {
-      FileSinkDesc fsd = writer;
-      writer = null;
-      return fsd;
-    }
-
-    public List<String> getPartColNames() {
-      return partColNames;
-    }
-
-    public void setPartColNames(List<String> partColNames) {
-      this.partColNames = partColNames;
-    }
-
-    public void setWriter(FileSinkDesc writer) {
-      this.writer = writer;
-    }
-
-    /**
-     * @return the isTemporary
-     */
-    @Explain(displayName = "isTemporary", displayOnlyOnTrue = true)
-    public boolean isTemporary() {
-      return isTemporary;
-    }
-
-    /**
-     * @return the isMaterialization
-     */
-    @Explain(displayName = "isMaterialization", displayOnlyOnTrue = true)
-    public boolean isMaterialization() {
-      return isMaterialization;
-    }
-    
-    public void setOwnerName(String ownerName) {
-      this.ownerName = ownerName;
-    }
-
-    public String getOwnerName() {
-      return this.ownerName;
     }
   }
 }
