+++ ql/src/java/org/apache/hadoop/hive/ql/ddl/database/drop/DropDatabaseAnalyzer.java
@@ -88,7 +88,7 @@ public class DropDatabaseAnalyzer extends BaseSemanticAnalyzer {
   }
 
   @Override
-  public boolean hasAcidResourcesInQuery() {
+  public boolean hasAcidResources() {
     // check DB tags once supported (i.e. ICEBERG_ONLY, ACID_ONLY, EXTERNAL_ONLY)
     return true;
   }
