+++ ql/src/java/org/apache/hadoop/hive/ql/lockmgr/DbTxnManager.java
@@ -24,7 +24,6 @@ import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hive.common.JavaUtils;
 import org.apache.hadoop.hive.common.ValidTxnList;
 import org.apache.hadoop.hive.common.ValidTxnWriteIdList;
-import org.apache.hadoop.hive.common.ValidWriteIdList;
 import org.apache.hadoop.hive.conf.Constants;
 import org.apache.hadoop.hive.conf.HiveConf;
 import org.apache.hadoop.hive.metastore.IMetaStoreClient;
@@ -329,7 +328,7 @@ public final class DbTxnManager extends HiveTxnManagerImpl {
    * @throws LockException
    */
   private void verifyState(QueryPlan queryPlan) throws LockException {
-    if(!isTxnOpen() && queryPlan.hasAcidResourcesInQuery()) {
+    if(!isTxnOpen() && queryPlan.hasAcidResources()) {
       throw new LockException("No transaction context for operation: " + queryPlan.getOperationName() +
         " for " + getQueryIdWaterMark(queryPlan));
     }
@@ -863,7 +862,7 @@ public final class DbTxnManager extends HiveTxnManagerImpl {
     }
     else if(!isExplicitTransaction) {
       assert numStatements == 1 : "numStatements=" + numStatements + " in implicit txn";
-      if (queryPlan.hasReadWriteAcidInQuery()) {
+      if (queryPlan.hasAcidReadWrite()) {
         //1st and only stmt in implicit txn and uses acid resource
         return true;
       }
